+++ pricing-ms/src/main/java/com/ecommerce/pricing/domain/repository/PriceRepository.java
@@ -6,5 +6,5 @@ import java.util.List;
 import com.ecommerce.pricing.domain.model.Price;
 
 public interface PriceRepository {
-    List<Price> findPrices(Long brandId, Long productId, LocalDateTime applicationDate);
+    List<Price> findPrices(Integer brandId, Long productId, LocalDateTime applicationDate);
 }
